Basic DataValue implementation

+++ include/parse.hpp
@@ -6,6 +6,6 @@
 using namespace std;
 
 void parseLine (string line, string& description, string& value);
-void parseExecution(string line, string& operation, float& operand_A, float&operand_B);
+void parseExecution(string line, string& operation, double& operand_A, double&operand_B);
 
 #endif
